Fix Symfony validation rule.
The API of Symfony\Component\Validator changed more than 2 years ago, so
this validation rule wasn't working for quite a long time.

This fixes the validator to work with versions >= 2.1 of
Symfony/Validation as the change on composer shows us. Although a bug
fix, this breaks compatibility with people already using this validator.

I was astonished to not find any tests for that validator also. :(
With those tests we can ensure that any change on the component API will
be noticed by us.