hhvm RecursiveTreeIterator support
This code is from the PHP source itself. It's required to run
Respect\Validation on hhvm, which lacks the class.

By adding it to the library folder, it's loaded automatically
by our testing routines.

Further documentation should be written, but only after a stable
build for hhvm appears on Travis.