MinimumAge test using Age rule.
This needs a fix, the "format" parameters accepted by "minimumAge"
rule is not supported by the new validator. It was supposed to be
compatible and do the exactly same thing as "minimumAge".

It would be cool to keep the same API as the "minimumAge", example:

	<?php
		// As it was before
		v::minimumAge(18, 'Y-m-d')->validate('1990-01-01'); // true

		// Needs to work that way
		v::age(18, 'Y-m-d')->validate('1990-01-01'); // true
		v::age(18, 65, 'Y-m-d')->validate('1990-01-01'); //true
	?>

It could work that way as the `format` will always be something
acceptable by `strtotime()` function. And an `age` (an small integer)
can also be a string (Ex: "18") but is not a valid format. So, if the
second parameter is given, and is not an `is_numeric()`, it is a format.