Remove Makefile from repository
The use of Respect\Foundation should be optional and not a dependency.
Since the "Makefile" is not part of the Respect\Validation library, I
see no point on keeping it into the repository but I add "Makefile" to
.gitignore list to allow people to use it.
Other related files/directories must in included into the global ignore
list.