Remove user-defined optional values
Since we have NotOptional rule now, we must keep the same opposite
behaviour between these two rules.

This commit also improve and increase tests from Optional rule and
updates the documentation.

No tests was made for messages on MODE_DEFAULT because it's not possible
for a while since we don't have such much granularity control, but after
some fixes on Not rule, maybe we will be able to do that, that's why I
kept the messages on OptionalException.