Validate month format using UNIX epoch
When using `date` rule to validate only the month portion of a date,
depending of *when* you validate something you might [have a strange
behavior][1].

If you want to validate `v::date('m')->validate('02')` for instance, if
you are on the 30th or 31th of a month you will get a `false` as an
answer. This happens because we use `DateTime::createFromFormat` and it
uses the current time for portions of the format which are omited.
So, the following happens internally:

    php > echo date('Y-m-d', strtotime('2015-02-30'));
    2015-03-02

As the month given by the `DateTime` object will be "03", the date check
now fails.
This fixes the problem by using `!` in the format (if it is only related
trying to validate a month) which forces the `createFromFormat` method
to use the Unix Epoch as the rest of the date instead of the current
time. Making this happen:

    php > echo date('Y-m-d', strtotime('1970-02-01'));
    1970-02-01

The test made will only break on 30th and 31th (and 29th depending on
the year), so it is not perfect.

    [1]: http://stackoverflow.com/questions/29346953