Rename method "validate()" to "isValid()"
Semantically it makes more sense that the method that returns whether
the input is valid or not for a specific rule be called "isValid()"
than "validate()". Also, in many languages, it's very common to prefix
methods that return boolean values in response to some logic with "is",
"can", "has", etc.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>