Remove PHP-CS-Fixer
It is not the first time that after a PHP-CS-Fixer upgrade our build
fails. Also, I've seemed a couple of files that have a wrong coding
standard, but PHP-CS-Fixer does not indicate failure for them.

This commit will remove the library as a dependency.

I plan to introduce PHP_CodeSniffer as soon as possible.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>