Use the last MINOR version in updaters
We have three workflows that automatically update the code. They are
using version 2.0 as a base to create their changes. We need to update
that because the last supported version is 2.1.

Since I do not think it is a good idea to change code every time a new
MINOR version is released, I moved that value to a repository secret.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>