Remove "setParams()" method from ValidationException
We already have a method called "updateParams()" which has a better name
for what it does. Besides, we don't really need this for this case.

Personally, I find it ugly to create a property in the rule just to have
it available in the exception, but that's how the design is.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>